backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
asttokens==3.0.0
    # via stack-data
click==8.1.8
    # via uvicorn
decorator==5.1.1
    # via
    #   ipdb
    #   ipython
django==5.1.6
    # via
    #   -r requirements.in
    #   django-braces
    #   django-cors-headers
    #   django-crispy-forms
    #   django-filter
    #   django-model-controller
    #   djangorestframework
django-braces==1.16.0
    # via django-model-controller
django-cors-headers==4.7.0
    # via -r requirements.in
django-crispy-forms==2.3
    # via django-model-controller
django-environ==0.12.0
    # via -r requirements.in
django-filter==25.1
    # via -r requirements.in
django-model-controller==0.4.6
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
executing==2.2.0
    # via stack-data
h11==0.14.0
    # via uvicorn
ipdb==0.13.13
    # via -r requirements.in
ipython==8.32.0
    # via ipdb
jedi==0.19.2
    # via ipython
matplotlib-inline==0.1.7
    # via ipython
numpy==2.2.3
    # via
    #   -r requirements.in
    #   pandas
    #   scipy
pandas==2.2.3
    # via -r requirements.in
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
prompt-toolkit==3.0.50
    # via ipython
psycopg2-binary==2.9.10
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pygments==2.19.1
    # via ipython
python-dateutil==2.9.0.post0
    # via pandas
pytz==2024.2
    # via
    #   -r requirements.in
    #   djangorestframework
    #   pandas
scipy==1.15.1
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via django
stack-data==0.6.3
    # via ipython
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
tzdata==2025.1
    # via pandas
uvicorn==0.32.1
    # via -r requirements.in
wcwidth==0.2.13
    # via prompt-toolkit
